Implement single endpoint for pages and posts pt1
This commit implements a portion of the work that needed to be
done for implementing a single rest endpoint for both page and post
slugs. Normally WP has separate endpoints for these and this will not
work when an url that is anything other than a slug is called.

Main part of the work is done in functions.php, where the endpoint will
live.