Implement router links, customrest endpoint usage
body content is retrieved using customrest endpoint.
React-router-dom Link and Navlink classes are used for the header and
the nav components

TODO:
FSA and dispatchMethod now accept a generic to distinguish their content
in this commit the generic is set to any all uses of these types. This needs
to be fixed