Implement Blog post retrieval, reorder actions
All page and post actions have been recollected under singular actions
as this is the way wp pushes the content to be handled.

At this stage the app is retrieving posts and pages with no issues,
there is a consistent data structure logic implemented. Actions and
Reducers are very organized.

Body and Blog retrieval logic is very lacking, this needs to be fixed
in the following commits