Implement Feature image and title for pages and posts
BodyContent component is created to separate Body's content
retrieval responsibility from the content display responsibility.

BodyContent compoent takes only the data that is meant to be
displayed on the screen. It accounts for the title and the
feature image of the page if either exists.

Styling has been updated to reflect the new behavior features
implemented.