Edit Navigaton, add current page ind, fix undefined slug
Navigation menu now completely relies on sass for styling, there
are no inline styles.

Also, there was an issue with the home page creating an undefined
slug as wp does not return a slug when homepage is called. This
has been fixed with the slug being run through an undefined
value check before being assigned to the link.

The mobile navigation menu now has a burger menu icon