Contact form restyling
Because of the new fields added in the contact form, the styling was
clashing with most viewports. This commit is intended to address
this issue.

Notes:
This commit may fall short in covering all the possible screen
shapes.
Some additional styling work will be needed once the form connects with
the rest api.