Update Singular retrieval types
This commit intends to create a better structure for singular retrieval
pattern from the server. There are still works do be done about the naming
of the types and interfaces. There are still problems with the elegance
of the code regarding slug retrieval and consequent decision to redirect
to native pages.