Rename elements to better fit their function
Body route and feature has been renamed as Slugged as the route and
the feature is responsible for fetch and display of items that are
accessible through slugs.

Body.component and its helpers have been renamed with their prior name
Canvas. This name may be prone change in the future if a better word is
found.