Refactor error codes
A general structure for managing error codes was partially implemented
with this commit.

What happens is that, every feature is going to have their own
set of error codes that will be nested inside an ERROR_CODES
object that is imported from the same module that is exported from
the rootreducer.

Warning:
Note that the current typechecking for ERROR_CODES type in the FSA
interface is faulty. Current type allows the ERROR_CODES object to be
used as a value while what we want is one of the values of the children
of the ERROR_CODES