Add tests, tidy types
An almost cookiecutter test model has been created for appfeature
functions.

A new folder named mock is created to hold the mock functions and
classes that may be needed for the testing of the components. With this
commit, 2 such fucntions have been added to the said module.

Error codes have been converted to string enums, with the value pattern
[feature name]/[error name]

New data types for refs have been created to ensure symmetry in IApp
interface.