Create config system
With this commit, environment variables are first parsed in a config
object

Some values that were previously handled as environment variables are
now received from window.config object that is populated by wordpress.

For now these include site title, description and separator. However,
the content of the config file can be
expanded in the future depending on theme needs.