Update styling to reflect changes by the client
The client has placed some content on the page. This has revealed some
use patterns and limitations due to the limited experience of the client
with the medium she is intending to use.

Chances are as follows:
The default in-page heading is now h2 instead of h3 as it was previously
intended. This means that the client will not have any access to the
font styling used in the hero section.

H2, H3, H4 have been moved up one level and a mock H4 has been create
to complete the heading list.

PostItem component headings have been moved from H4 to H3

Hero section stylish font has been moved to a mixin

Many grid-template area definitions have been changed to reflect the
H{x} change.

.wp-block-group styling has been specified as the immediate child (>) of
the parent to allow the client to add multiple paragaphs in the home
page.

The block that contains the paragraphs is defined as the next tag after
h2 in every section.