Update logos, change extension
Per client's request for increasing spacing in the letters of the
logo/typography, the logo file has been updated.

Logo's extension has been switched from svg to png for achieving a
smaller file size. For this, certain css styling needed to be revised.

It should be noted that the finer details of image sizing are still to
be done.

This commit also sees a sizing adjustment for the hero title. Now
the mixin responsible for the sizing starts sizing from 250px until
1920px.