Update scripts
With it becoming clear that setting up jenkins for this project will
take a bit longer than what the deadline for the project will allow,
the best course of action was to update the scripts used in the repo to
be modular enough to be used both standalone and with jenkins in the
future.

For this, a couple of changes were made to the scripts which will be
outlined below:

1- environment variables are introduced to manage the vm instance and the
directories involved to be controlled from a single file.

2- There were slight changes with the docker-compose files for dev, test
and production due to the script path changes.

3- Prod_init file that was introduced before has been altered to
actually init the production machine. A new script file prod_update.sh
is introduced with the aim to update certain sections of the codebase
such as the theme or the docker config file.

4- test_up file has been cleared of some lines that did nothing for the
current iteration of the file while now including  a new line for getting
rid of the useless files that the yarn build process creates.

5- .env for the react app has received a new variable to stop cra from
creating sourcemaps for production

6- scripts of production and dev of basakbeykoz_db has been completely
merged. Now a single folder serves both environments

7- docker-compose file for dev no longer has the scripts volume mounted
on the wp service as there was no script to be used for this service.

8- changes were made on docker-compose filed for dev, test and
production to reflect the changes with scripts for db being merged
to be the same ones for all environments