Update homepage styling
Footer artwork received a bit of a margin on its left and its right
through width adjustment that created some clearance on its sides.
Which looks much better on smaller mobile devices.

ul got a base level padding adjustment to make it mre condensed on
mobile screens.

home page sections had their max height restrictions on larger screens
lifted as the client wants to post a lot of words in each of these.

the block images in sections :not(first) now have a min height property
is proportional to the width of the screen. This is only valid until
svga screen width. This was done as the height restrictions being lifted
allows more more space.

Grid adjustments were made in multiple home page sections to make them
work better with the new auto height of the sections