Fix: Footer padding on pges other than home
In pages other than home, the footer was overlapping with the content.
This commit fixes that issue. Note that the footer padding is not
responsive in this commit. It could be useful to switch to a
responsive value in a future commit if deemed necessary.

This commit also sees a minor change in the Social.view component to
increase readability.