Design: add homepage star
The client was requesting a star behind the h2 of one of the homepage
sections.

The star has been placed with responsive size that ties the size and
position of the star to h2 line-height.

This commit also refactors get-typography method in mixins to
make it easier to retrieve responsive sizes of font and line-height