Refactor, Feature: ContactFormApi
Slugged.feature code has been refactored to remove the injections to
their own modules.

Contact form api has been implemented and is listening to user
actions and talking to the server in this commit. However, error
handling has not been implemented yet.

The 2000msec timeout for injecting page functions has been replaced
with process.nextTick(() at the level of the method 'getSlugOnLoad'.
Meaning that the injection method does not deal with the timing
of the injection anymore.