Feature: implement share
Share feture was implemented with:
facebook, twitter, whatsapp and linkedin.

Currently the styling is a bit problematic. And the component is
deliberately set to be hidden in small screens. This can be changed if the
decision is made to keep the share items.

A magic number was discovered in the feature image component of canvas. The
image is rendering -5px to the left of the viewport. This wasn't fixed
in this commit.

For the share component to work as expected, excerpt value was added to
the chain of types that start from SingularItemForView till
Canvas.layout. Relevant alterations for this change was made on every
module that typescript warned about. This includes some routes,
tests, and features.

Note that some routes such as blog and 404 have excerpt values that
were fudged in this commit. These shall be properly retrieved from php
through a ui that the client has access to.

Feature image now has a bit of a filter to make it fit better with the
colors theme of the page.