Design: implement padding consistency
There was some inconsistencies related to page whitespace padding,
padding of figure and similar framed components.

These paddign amounts have been tied to 2 different sets of
component properties.

This commit also contains assignment of margins to heading tags, and
exclusion of the said margins from homepage and canvas components. The
reason for setting margin to headings is to make the page and post
content  that will be written by the cllient to be easily managed.

There were a few locations where the pattern var(--gap) was still
used. These have been replaced with get-gap().