Design: Fix minor issues
Some <a> elements lacked margin declarations, which could cause
alignment issues in some browsers. This was fixed

The decoration next to the logo was sized to single gap while in
previous commits it was sized at 2. This was returned to its previous
value.

A pointless z-index declaration in home_1 was removed as it was already
overwritten with another z-index a few lines down.