Feature: add shareDisabled property for active
While the app was waiting for the very first content to load, the share
buttons were showing on the background of the semi-transparent spinner
screen. This visual was not appealing.

This commit introduces a shareDisabled propery in the active object in
the app store. True value for this propery results in
CanvasView avoiding to display the share screen.

This is a stop-gap measure that shall be replaced by a proper
"app is loading" state which avoids displaying the canvas before
the app is ready.