Feature, Design: Fix homepage post blocks
The design of post blocks on the frontpage and their functionality
has been improved.

A function that was added to functions.php in prior commits for
controlling properties of post excerpts has been replaced with a more
comprehensive function that handles tag based controls, exclusion of
scripts (if there may be any) and it limits the length of the
excerpts to 50 words.

Injections that are made by home.injection.ts have been improved

The "Similar Reads" h3 title in the canvas has been replaced with its
turkish translation

.wp-block-latest-posts  styling has been divided into separate modules
to ease the task of styling the component.

Several adjustments were made after the module refactoring:

A new typography size called p-small was introduced. This is meant to be
used inside the post cards and probably other areas where smaller
paragraphs could be useful.

margin-bottom for p and p-small, which is a smaller p element that is
used for the cards now have a margin bottom that is equal to their
line height. This was done to make it easier to predict the height of
content. The reason for this will be made clear later.

limit-paragraph-lines mixin now allows selection of the text type such
as p, h4 and alike. This was done due to the need for limiting content
height for things other than p.

wp-posts was being imported by many modules that had no use for this
module. These have been cleaned up.

home_4 and home_5 received some updates regarding the size of the
post card dimensions. Because of the nature of the content that is on
the page, it became necessary to define the sizes of post card
contents top down.

Because of this, these two modules now contain
some variables and subsequent variable calculations that specify
the line count of card components such as the text and the title,
the feature image height, padding amount in the card, etc.

These specifications are later converted to calc values in order to
set a height for the card container. These values are also given to
the wp-posts mixin to be used for the specified properties.

wp-posts modules have also seen some simplification in terms of the
parameters used. Before, whether the content would be displayed as
row or column was given separately, as well as where the title element
would have been placed on top of the card or bottom as the author name.
Now these decisions are based on the type of the content; whether
it is posts, or success. This change has been done as it became clear
that the use for these mixins will not get any more complicated beyond
supporting 2 different types.

Cards now have a hover state in which the color of the card changes
depending on what color has been given by the respective homepage
section module.

Some documentation has been written a lot more has to be done on
this front.