Feature: retrieve social items from php
- Previously social items such as the instagram and linkedin links were
a part of the state.

However this content does not change and it
always made more sense to retrieve these items from php as the user
can define these values themselves.

This commit makes all the changes necessary to move the social values
from state to php. List of things altered are:

+ index.php
+ global types -> window.config
+ config module
+ features -> social
+ views -> social